TST: move fixture to conftest
