+++ bluesky/tests/conftest.py
@@ -2,6 +2,12 @@ import asyncio
 from bluesky.run_engine import RunEngine
 from bluesky.examples import Mover, SynGauss
 import pytest
+from types import SimpleNamespace
+from ophyd.pseudopos import (PseudoPositioner, PseudoSingle,
+                             real_position_argument, pseudo_position_argument)
+from ophyd.positioner import SoftPositioner
+from ophyd.signal import Signal
+from ophyd import (Component as C)
 
 
 @pytest.fixture(scope='function')
@@ -40,3 +46,56 @@ def db(request):
     from databroker.tests.utils import build_sqlite_backed_broker
     db = build_sqlite_backed_broker(request)
     return db
+
+
+@pytest.fixture()
+def hw():
+    class SPseudo3x3(PseudoPositioner):
+        pseudo1 = C(PseudoSingle, limits=(-10, 10), egu='a')
+        pseudo2 = C(PseudoSingle, limits=(-10, 10), egu='b')
+        pseudo3 = C(PseudoSingle, limits=None, egu='c')
+        real1 = C(SoftPositioner, init_pos=0)
+        real2 = C(SoftPositioner, init_pos=0)
+        real3 = C(SoftPositioner, init_pos=0)
+
+        sig = C(Signal, value=0)
+
+        @pseudo_position_argument
+        def forward(self, pseudo_pos):
+            pseudo_pos = self.PseudoPosition(*pseudo_pos)
+            # logger.debug('forward %s', pseudo_pos)
+            return self.RealPosition(real1=-pseudo_pos.pseudo1,
+                                     real2=-pseudo_pos.pseudo2,
+                                     real3=-pseudo_pos.pseudo3)
+
+        @real_position_argument
+        def inverse(self, real_pos):
+            real_pos = self.RealPosition(*real_pos)
+            # logger.debug('inverse %s', real_pos)
+            return self.PseudoPosition(pseudo1=-real_pos.real1,
+                                       pseudo2=-real_pos.real2,
+                                       pseudo3=-real_pos.real3)
+
+    class SPseudo1x3(PseudoPositioner):
+        pseudo1 = C(PseudoSingle, limits=(-10, 10))
+        real1 = C(SoftPositioner, init_pos=0)
+        real2 = C(SoftPositioner, init_pos=0)
+        real3 = C(SoftPositioner, init_pos=0)
+
+        @pseudo_position_argument
+        def forward(self, pseudo_pos):
+            pseudo_pos = self.PseudoPosition(*pseudo_pos)
+            # logger.debug('forward %s', pseudo_pos)
+            return self.RealPosition(real1=-pseudo_pos.pseudo1,
+                                     real2=-pseudo_pos.pseudo1,
+                                     real3=-pseudo_pos.pseudo1)
+
+        @real_position_argument
+        def inverse(self, real_pos):
+            real_pos = self.RealPosition(*real_pos)
+            # logger.debug('inverse %s', real_pos)
+            return self.PseudoPosition(pseudo1=-real_pos.real1)
+
+    return SimpleNamespace(pseudo3x3=SPseudo3x3(name='pseudo3x3'),
+                           pseudo1x3=SPseudo1x3(name='pseudo1x3'),
+                           sig=Signal(name='sig', value=0))
