+++ bluesky/tests/test_utils.py
@@ -1,15 +1,10 @@
 import pytest
-from types import SimpleNamespace
+
 from functools import reduce
 import operator
 
 from bluesky.utils import ensure_generator, Msg, merge_cycler
 from cycler import cycler
-from ophyd.pseudopos import (PseudoPositioner, PseudoSingle,
-                             real_position_argument, pseudo_position_argument)
-from ophyd.positioner import SoftPositioner
-from ophyd.signal import Signal
-from ophyd import (Component as C)
 
 
 def test_single_msg_to_gen():
@@ -19,59 +14,6 @@ def test_single_msg_to_gen():
 
     assert len(m_list) == 1
     assert m_list[0] == m
-
-
-@pytest.fixture()
-def hw():
-    class SPseudo3x3(PseudoPositioner):
-        pseudo1 = C(PseudoSingle, limits=(-10, 10), egu='a')
-        pseudo2 = C(PseudoSingle, limits=(-10, 10), egu='b')
-        pseudo3 = C(PseudoSingle, limits=None, egu='c')
-        real1 = C(SoftPositioner, init_pos=0)
-        real2 = C(SoftPositioner, init_pos=0)
-        real3 = C(SoftPositioner, init_pos=0)
-
-        sig = C(Signal, value=0)
-
-        @pseudo_position_argument
-        def forward(self, pseudo_pos):
-            pseudo_pos = self.PseudoPosition(*pseudo_pos)
-            # logger.debug('forward %s', pseudo_pos)
-            return self.RealPosition(real1=-pseudo_pos.pseudo1,
-                                     real2=-pseudo_pos.pseudo2,
-                                     real3=-pseudo_pos.pseudo3)
-
-        @real_position_argument
-        def inverse(self, real_pos):
-            real_pos = self.RealPosition(*real_pos)
-            # logger.debug('inverse %s', real_pos)
-            return self.PseudoPosition(pseudo1=-real_pos.real1,
-                                       pseudo2=-real_pos.real2,
-                                       pseudo3=-real_pos.real3)
-
-    class SPseudo1x3(PseudoPositioner):
-        pseudo1 = C(PseudoSingle, limits=(-10, 10))
-        real1 = C(SoftPositioner, init_pos=0)
-        real2 = C(SoftPositioner, init_pos=0)
-        real3 = C(SoftPositioner, init_pos=0)
-
-        @pseudo_position_argument
-        def forward(self, pseudo_pos):
-            pseudo_pos = self.PseudoPosition(*pseudo_pos)
-            # logger.debug('forward %s', pseudo_pos)
-            return self.RealPosition(real1=-pseudo_pos.pseudo1,
-                                     real2=-pseudo_pos.pseudo1,
-                                     real3=-pseudo_pos.pseudo1)
-
-        @real_position_argument
-        def inverse(self, real_pos):
-            real_pos = self.RealPosition(*real_pos)
-            # logger.debug('inverse %s', real_pos)
-            return self.PseudoPosition(pseudo1=-real_pos.real1)
-
-    return SimpleNamespace(pseudo3x3=SPseudo3x3(name='pseudo3x3'),
-                           pseudo1x3=SPseudo1x3(name='pseudo1x3'),
-                           sig=Signal(name='sig', value=0))
 
 
 @pytest.mark.parametrize('traj',
